Restructured plots, and made supercell optional in dashboard
