+++ src/gemdat/dashboard/run.py
@@ -71,24 +71,26 @@ with st.sidebar:
     sites_filename = str(
         st.selectbox('Load sites from known material', KNOWN_MATERIALS))
 
-    st.markdown('Supercell (x,y,z)')
-    col1, col2, col3 = st.columns(3)
-    supercell = (int(
-        col1.number_input('supercell x',
-                          min_value=1,
-                          value=1,
-                          label_visibility='collapsed',
-                          help=None)),
-                 int(
-                     col2.number_input('supercell y',
-                                       min_value=1,
-                                       value=1,
-                                       label_visibility='collapsed')),
-                 int(
-                     col3.number_input('supercell z',
-                                       min_value=1,
-                                       value=1,
-                                       label_visibility='collapsed')))
+    manual_supercell = st.checkbox('Manual Supercell')
+    if manual_supercell:
+        st.markdown('Supercell (x,y,z)')
+        col1, col2, col3 = st.columns(3)
+        supercell = (int(
+            col1.number_input('supercell x',
+                              min_value=1,
+                              value=1,
+                              label_visibility='collapsed',
+                              help=None)),
+                     int(
+                         col2.number_input('supercell y',
+                                           min_value=1,
+                                           value=1,
+                                           label_visibility='collapsed')),
+                     int(
+                         col3.number_input('supercell z',
+                                           min_value=1,
+                                           value=1,
+                                           label_visibility='collapsed')))
 
     st.markdown('## Enable Error Analysis')
     do_error = st.checkbox('Error Analysis')
@@ -136,7 +138,14 @@ with col3:
 
 tab1, tab2, tab3 = st.tabs(['Default plots', 'RDF plots', 'Density plots'])
 
-sites_structure = load_known_material(sites_filename, supercell=supercell)
+if manual_supercell:
+    sites_structure = load_known_material(sites_filename, supercell=supercell)
+else:
+    sites_structure = load_known_material(sites_filename)
+    zipped_parameters = zip(trajectory.get_lattice().abc,
+                            sites_structure.lattice.parameters)
+    supercell = [round(a / b) for a, b in zipped_parameters]  # type: ignore
+    sites_structure.make_supercell(supercell)
 
 if not is_lattice_similar(trajectory.get_lattice(), sites_structure):
     st.error('Lattices are not similar!')
@@ -180,14 +189,14 @@ with tab1:
     diff_trajectory = trajectory.filter(diffusing_element)
 
     figures = (
-        plots.displacement_per_element2(trajectory=trajectory),
+        plots.displacement_per_element(trajectory=trajectory),
         plots.displacement_per_site(trajectory=diff_trajectory),
-        plots.displacement_histogram2(trajectory=trajectory, n_parts=n_parts),
+        plots.displacement_histogram(trajectory=trajectory, n_parts=n_parts),
         plots.frequency_vs_occurence(trajectory=diff_trajectory),
-        plots.vibrational_amplitudes2(trajectory=diff_trajectory,
-                                      n_parts=n_parts),
-        plots.jumps_vs_distance2(sites=sites, n_parts=n_parts),
-        plots.jumps_vs_time2(sites=sites, n_parts=n_parts),
+        plots.vibrational_amplitudes(trajectory=diff_trajectory,
+                                     n_parts=n_parts),
+        plots.jumps_vs_distance(sites=sites, n_parts=n_parts),
+        plots.jumps_vs_time(sites=sites, n_parts=n_parts),
         plots.collective_jumps(sites=sites),
         plots.jumps_3d(sites=sites),
     )
