+++ src/gemdat/plots/__init__.py
@@ -1,44 +1,37 @@
 """This module contains all the plots that Gemdat can generate."""
 from __future__ import annotations
 
-from ._density import density
-from ._displacements import (
-    displacement_histogram,
-    displacement_histogram2,
-    displacement_per_element,
-    displacement_per_element2,
-    displacement_per_site,
-)
-from ._jumps import (
+# Matplotlib plots
+from .matplotlib._displacements import displacement_per_site
+from .matplotlib._jumps import (
     collective_jumps,
     jumps_3d,
     jumps_3d_animation,
+)
+from .matplotlib._rdf import radial_distribution
+from .matplotlib._vibration import frequency_vs_occurence
+
+# Plotly plots (matplotlib version might be available)
+from .plotly._density import density
+from .plotly._displacements import (
+    displacement_histogram,
+    displacement_per_element,
+)
+from .plotly._jumps import (
     jumps_vs_distance,
-    jumps_vs_distance2,
     jumps_vs_time,
-    jumps_vs_time2,
 )
-from ._rdf import radial_distribution
-from ._vibration import (
-    frequency_vs_occurence,
-    vibrational_amplitudes,
-    vibrational_amplitudes2,
-)
+from .plotly._vibration import vibrational_amplitudes
 
 __all__ = [
     'density',
     'displacement_per_site',
     'displacement_per_element',
-    'displacement_per_element2',
     'displacement_histogram',
-    'displacement_histogram2',
     'frequency_vs_occurence',
     'vibrational_amplitudes',
-    'vibrational_amplitudes2',
     'jumps_vs_distance',
-    'jumps_vs_distance2',
     'jumps_vs_time',
-    'jumps_vs_time2',
     'collective_jumps',
     'jumps_3d',
     'jumps_3d_animation',
